chrome/test/data/xr/e2e_test_files/html/test_content_resizing_does_not_affect_webpage.html

<!doctype html>
<!--
Tests that resizing the VR Browser content quad does not resize the dimensions
that are reported to the webpage.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script>
      setup({single_test: true});

      var initialWidth = 0;
      var initialHeight = 0;
      function stepGetInitialDimensions() {
        initialWidth = window.innerWidth;
        initialHeight = window.innerHeight
        finishJavaScriptStep();
      }

      function stepCheckDimensionsAfterResize() {
        assert_equals(window.innerWidth, initialWidth, 'Resize affected width');
        assert_equals(window.innerHeight, initialHeight, 'Resize affected height');
        done();
      }
    </script>
  </body>
</html>
